TS: split RelatedAdvanced from components/Related
