+++ src/components/Related/index.js
@@ -11,122 +11,13 @@ import Loading from 'components/SimpleCommonComponents/Loading';
 
 import { toPlural } from 'utils/pages';
 
-import RelatedTable from './RelatedTable';
-// $FlowFixMe
-import EntriesOnStructure from './DomainEntriesOnStructure';
 // $FlowFixMe
 import RelatedSimple from './RelatedSimple';
 // $FlowFixMe
-import { findIn, filterIn } from 'utils/processDescription/filterFuncions';
-import StructureOnProtein from './DomainStructureOnProtein';
+import RelatedAdvanced from './RelatedAdvanced';
+// $FlowFixMe
+import { findIn } from 'utils/processDescription/filterFuncions';
 import { getUrlForMeta, getReversedUrl } from 'higherOrder/loadData/defaults';
-
-import { foundationPartial } from 'styles/foundation';
-
-import ebiGlobalStyles from 'ebi-framework/css/ebi-global.css';
-
-const f = foundationPartial(ebiGlobalStyles);
-
-/*:: type relatedAdvancedProps = {
-  mainData: Object,
-  secondaryData: Array<Object>,
-  isStale: boolean,
-  mainType: string,
-  focusType: string,
-  actualSize: number,
-  otherFilters?: Array<Object>,
-  dataBase: {
-    payload: Object,
-    loading: boolean
-  },
-  secondaryDataLoading: boolean,
-}; */
-/*:: type state = {
-  showTaxoInfo: boolean,
-}; */
-export class _RelatedAdvanced extends PureComponent /*:: <relatedAdvancedProps> */ {
-  static propTypes = {
-    mainData: T.object.isRequired,
-    secondaryData: T.arrayOf(T.object).isRequired,
-    isStale: T.bool.isRequired,
-    mainType: T.string.isRequired,
-    focusType: T.string.isRequired,
-    actualSize: T.number,
-    otherFilters: T.array,
-    dataBase: T.shape({
-      payload: T.object,
-      loading: T.bool.isRequired,
-    }).isRequired,
-    secondaryDataLoading: T.bool.isRequired,
-  };
-
-  render() {
-    const {
-      mainData,
-      secondaryData,
-      isStale,
-      mainType,
-      focusType,
-      actualSize,
-      otherFilters,
-      dataBase,
-      secondaryDataLoading,
-    } = this.props;
-    return (
-      <div className={f('row', 'column')}>
-        {secondaryDataLoading ? (
-          <Loading />
-        ) : (
-          <div>
-            {mainType === 'protein' && focusType === 'structure' ? (
-              <StructureOnProtein
-                structures={secondaryData}
-                protein={mainData}
-              />
-            ) : null}
-            {mainType === 'structure' && focusType === 'entry' ? (
-              <EntriesOnStructure entries={secondaryData} />
-            ) : null}
-            <RelatedTable
-              mainType={mainType}
-              mainData={mainData}
-              secondaryData={secondaryData}
-              focusType={focusType}
-              otherFilters={otherFilters}
-              dataBase={dataBase}
-              isStale={isStale}
-              actualSize={actualSize}
-              otherProps={{ ...this.props }}
-            />
-          </div>
-        )}
-      </div>
-    );
-  }
-}
-
-const mapStateToPropsAdvanced = createSelector(
-  (state) => state.customLocation.description.main.key,
-  (state) =>
-    findIn(
-      state.customLocation.description,
-      (value /*: {isFilter:boolean, order:number} */) =>
-        value.isFilter && value.order === 1,
-    ),
-  (state) =>
-    filterIn(
-      state.customLocation.description,
-      (value /*: {isFilter:boolean, order:number} */) =>
-        value.isFilter && value.order !== 1,
-    ),
-  (mainType, [focusType, { db: focusDB }], otherFilters) => ({
-    mainType,
-    focusType,
-    focusDB,
-    otherFilters,
-  }),
-);
-const RelatedAdvanced = connect(mapStateToPropsAdvanced)(_RelatedAdvanced);
 
 const mapStateToPropsAdvancedQuery = createSelector(
   (state) => state.customLocation.description.main.key,
